remove module inheritance
