+++ src/careamics/lightning/dataset_ng/lightning_modules/n2v_module.py
@@ -1,21 +1,27 @@
-"""Noise2Void Lightning DataModule."""
+"""Noise2Void Lightning Module."""
 
-from typing import Any, Union
+from typing import Any
 
-from careamics.config import (
-    N2VAlgorithm,
-)
+import pytorch_lightning as L
+import torch
+from torch import nn
+from torchmetrics import MetricCollection
+from torchmetrics.image import PeakSignalNoiseRatio
+
+from careamics.config import N2VAlgorithm, algorithm_factory
 from careamics.dataset_ng.dataset import ImageRegionData
 from careamics.losses import n2v_loss
+from careamics.models.unet import UNet
 from careamics.transforms import N2VManipulateTorch
 from careamics.utils.logging import get_logger
+from careamics.utils.torch_utils import get_optimizer, get_scheduler
 
-from .unet_module import UnetModule
+from .module_utils import log_training_stats, log_validation_stats
 
 logger = get_logger(__name__)
 
 
-class N2VModule(UnetModule):
+class N2VModule(L.LightningModule):
     """CAREamics PyTorch Lightning module for N2V algorithm.
 
     Parameters
@@ -25,8 +31,8 @@ class N2VModule(UnetModule):
         dictionary.
     """
 
-    def __init__(self, algorithm_config: Union[N2VAlgorithm, dict]) -> None:
-        """Instantiate N2V DataModule.
+    def __init__(self, algorithm_config: N2VAlgorithm | dict) -> None:
+        """Instantiate N2V Module.
 
         Parameters
         ----------
@@ -34,43 +40,55 @@ class N2VModule(UnetModule):
             Configuration for the N2V algorithm, either as an N2VAlgorithm instance or a
             dictionary.
         """
-        super().__init__(algorithm_config)
+        super().__init__()
 
-        assert isinstance(
-            algorithm_config, N2VAlgorithm
-        ), "algorithm_config must be a N2VAlgorithm"
+        if isinstance(algorithm_config, dict):
+            algorithm_config = algorithm_factory(algorithm_config)
+        if not isinstance(algorithm_config, N2VAlgorithm):
+            raise TypeError("algorithm_config must be a N2VAlgorithm")
 
+        self.config = algorithm_config
+        self.model: nn.Module = UNet(**algorithm_config.model.model_dump())
         self.n2v_manipulate = N2VManipulateTorch(
             n2v_manipulate_config=algorithm_config.n2v_config
         )
         self.loss_func = n2v_loss
 
-    def _load_best_checkpoint(self) -> None:
-        """Load the best checkpoint for N2V model."""
-        logger.warning(
-            "Loading best checkpoint for N2V model. Note that for N2V, "
-            "the checkpoint with the best validation metrics may not necessarily "
-            "have the best denoising performance."
-        )
-        super()._load_best_checkpoint()
+        self.metrics = MetricCollection(PeakSignalNoiseRatio())
+
+    def forward(self, x: torch.Tensor) -> torch.Tensor:
+        """
+        Forward pass.
+
+        Parameters
+        ----------
+        x : torch.Tensor
+            Input tensor.
+
+        Returns
+        -------
+        torch.Tensor
+            Model output tensor.
+        """
+        return self.model(x)
 
     def training_step(
         self,
-        batch: Union[tuple[ImageRegionData], tuple[ImageRegionData, ImageRegionData]],
-        batch_idx: Any,
-    ) -> Any:
+        batch: tuple[ImageRegionData] | tuple[ImageRegionData, ImageRegionData],
+        batch_idx: int,
+    ) -> torch.Tensor:
         """Training step for N2V model.
 
         Parameters
         ----------
         batch : ImageRegionData or (ImageRegionData, ImageRegionData)
             A tuple containing the input data and the target data.
-        batch_idx : Any
+        batch_idx : int
             The index of the current batch in the training loop.
 
         Returns
         -------
-        Any
+        torch.Tensor
             The loss value for the current training step.
         """
         x = batch[0]
@@ -78,14 +96,14 @@ class N2VModule(UnetModule):
         prediction = self.model(x_masked)
         loss = self.loss_func(prediction, x_original, mask)
 
-        self._log_training_stats(loss, batch_size=x.data.shape[0])
+        log_training_stats(self, loss, batch_size=x.data.shape[0])
 
         return loss
 
     def validation_step(
         self,
-        batch: Union[tuple[ImageRegionData], tuple[ImageRegionData, ImageRegionData]],
-        batch_idx: Any,
+        batch: tuple[ImageRegionData] | tuple[ImageRegionData, ImageRegionData],
+        batch_idx: int,
     ) -> None:
         """Validation step for N2V model.
 
@@ -93,8 +111,8 @@ class N2VModule(UnetModule):
         ----------
         batch : ImageRegionData or (ImageRegionData, ImageRegionData)
             A tuple containing the input data and the target data.
-        batch_idx : Any
-            The index of the current batch in the training loop.
+        batch_idx : int
+            The index of the current batch in the validation loop.
         """
         x = batch[0]
 
@@ -103,4 +121,65 @@ class N2VModule(UnetModule):
 
         val_loss = self.loss_func(prediction, x_original, mask)
         self.metrics(prediction, x_original)
-        self._log_validation_stats(val_loss, batch_size=x.data.shape[0])
+        log_validation_stats(
+            self, val_loss, batch_size=x.data.shape[0], metrics=self.metrics
+        )
+
+    def predict_step(
+        self,
+        batch: tuple[ImageRegionData] | tuple[ImageRegionData, ImageRegionData],
+        batch_idx: int,
+    ) -> ImageRegionData:
+        """Prediction step for N2V model.
+
+        Parameters
+        ----------
+        batch : ImageRegionData or (ImageRegionData, ImageRegionData)
+            A tuple containing the input data and optionally the target data.
+        batch_idx : int
+            The index of the current batch in the prediction loop.
+
+        Returns
+        -------
+        ImageRegionData
+            The output batch containing the predictions.
+        """
+        x = batch[0]
+        # TODO: add TTA
+        prediction = self.model(x.data)
+
+        normalization = self._trainer.datamodule.predict_dataset.normalization
+        denormalized_output = normalization.denormalize(prediction).cpu().numpy()
+
+        output_batch = ImageRegionData(
+            data=denormalized_output,
+            source=x.source,
+            data_shape=x.data_shape,
+            dtype=x.dtype,
+            axes=x.axes,
+            region_spec=x.region_spec,
+            additional_metadata={},
+        )
+        return output_batch
+
+    def configure_optimizers(self) -> dict[str, Any]:
+        """Configure optimizer and learning rate scheduler.
+
+        Returns
+        -------
+        dict[str, Any]
+            A dictionary containing the optimizer and learning rate scheduler.
+        """
+        optimizer_func = get_optimizer(self.config.optimizer.name)
+        optimizer = optimizer_func(
+            self.model.parameters(), **self.config.optimizer.parameters
+        )
+
+        scheduler_func = get_scheduler(self.config.lr_scheduler.name)
+        scheduler = scheduler_func(optimizer, **self.config.lr_scheduler.parameters)
+
+        return {
+            "optimizer": optimizer,
+            "lr_scheduler": scheduler,
+            "monitor": "val_loss",
+        }
